got 25 points

+++ GridlandMetro.h
@@ -9,5 +9,5 @@ using namespace std;
 class GridlandMetro
 {
 public:
-	int gridlandMetro(int n, int m, int k, vector<vector<int>> track);
+	long long gridlandMetro(int n, int m, int k, vector<vector<int>> track);
 };
